requirements.txt
(ee656) adisharmaruda@gandiva:~/ee656/AirCompressor_Data/ee656_course_project$ conda list --export
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.0
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.9.0=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.4.26=hbd8a1cb_0
category-encoders=2.8.1=pypi_0
certifi=2025.4.26=pypi_0
charset-normalizer=3.4.2=pypi_0
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.11=py310h6a678d5_0
decorator=5.2.1=pyhd8ed1ab_0
entrypoints=0.4=pyhd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
filelock=3.13.1=pypi_0
fonttools=4.58.1=pypi_0
fsspec=2024.6.1=pypi_0
grpcio=1.72.1=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.1.3=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.33.0=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.37.0=pyh8f84b5b_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.4=pypi_0
joblib=1.5.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.8.1=pyh31011fe_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown=3.8=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mpmath=1.3.0=pypi_0
mrmr-selection=0.2.8=pypi_0
narwhals=1.41.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.3=pypi_0
numpy=2.1.2=pypi_0
nvidia-cublas-cu11=11.11.3.6=pypi_0
nvidia-cuda-cupti-cu11=11.8.87=pypi_0
nvidia-cuda-nvrtc-cu11=11.8.89=pypi_0
nvidia-cuda-runtime-cu11=11.8.89=pypi_0
nvidia-cudnn-cu11=9.1.0.70=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.3.0.86=pypi_0
nvidia-cusolver-cu11=11.4.1.48=pypi_0
nvidia-cusparse-cu11=11.7.5.86=pypi_0
nvidia-nccl-cu11=2.21.5=pypi_0
nvidia-nvtx-cu11=11.8.86=pypi_0
opencv-python=4.11.0.86=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.2.3=pypi_0
parso=0.8.4=pyhd8ed1ab_1
patsy=1.0.1=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.0.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pyhe01879c_0
plotly=6.1.2=pypi_0
polars=1.31.0=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
protobuf=6.31.1=pypi_0
psutil=7.0.0=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
py-cpuinfo=9.0.0=pypi_0
pycocotools=2.0.10=pypi_0
pydantic=2.11.5=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.1=pyhd8ed1ab_0
pyparsing=3.2.3=pypi_0
python=3.10.16=he870216_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.10=2_cp310
pytz=2025.2=pypi_0
pywavelets=1.8.0=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=26.2.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
safetensors=0.5.3=pypi_0
sarvamai=0.1.5=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.7.0=pypi_0
scipy=1.15.3=pypi_0
seaborn=0.13.2=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.3=pyhd8ed1ab_1
statsmodels=0.14.4=pypi_0
sympy=1.13.3=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.5.10=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.1=pypi_0
torch=2.7.0+cu118=pypi_0
torchaudio=2.7.0+cu118=pypi_0
torchvision=0.22.0+cu118=pypi_0
tornado=6.1=py310h5764c6d_3
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.52.4=pypi_0
triton=3.3.0=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.14.0=pyhe01879c_0
tzdata=2025.2=pypi_0
ultralytics=8.3.146=pypi_0
ultralytics-thop=2.0.14=pypi_0
urllib3=2.4.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_1
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xz=5.6.4=h5eee18b_1
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1